requirements-dev.txt
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.1.0
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
